refactor: break up module graph implementation more granularly
