+++ assembly/src/assembler/module_graph/analysis/mod.rs
@@ -1,0 +1,3 @@
+mod rewrite_check;
+
+pub use self::rewrite_check::MaybeRewriteCheck;
