+++ assembly/src/assembler/module_graph/rewrites/mod.rs
@@ -1,0 +1,3 @@
+mod module;
+
+pub use self::module::ModuleRewriter;
